experiments/LegacyExperiments/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
exceptiongroup==1.3.0
google-auth==2.41.1
google-genai==1.39.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
loguru==0.7.3
mcp==1.15.0
numpy==2.2.6
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
rdkit==2025.3.6
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
